Add files via upload
It includes basic CRUD operations for customer data, a login form UI, and integration with an in-memory H2 database.I used Spring Data JPA for persistence and Thymeleaf for rendering the login page. While it doesn't have full authentication yet, it's structured in a way that allows for easy extension — like adding Spring Security, user roles, and transaction features later.